Feature: Plugin Path & Refactor
