+++ .vscode/settings.json
@@ -1,3 +1,3 @@
 {
-  "cSpell.words": ["Bucsy", "falkor", "pandoc", "theonethread", "npmjs", "deduped", "snyk", "codeql"]
+  "cSpell.words": ["Bucsy", "falkor", "pandoc", "theonethread", "npmjs", "deduped", "snyk", "codeql", "ifnset"]
 }
