+++ man/man.md
@@ -9,8 +9,8 @@
 # SYNOPSIS
 
 ```
-falkor-commander [(--scope <scope>)] [(--keyword <keyword>)] [<tasks>...] [(-- <answers>...)]
-falkor-commander [(--s <scope>)] [(--k <keyword>)] [<tasks>...] [(-- <answers>...)]
+falkor-commander [(--scope <scope>) | (--path <path>)] [(--keyword <keyword>)] [<tasks>...] [(-- <answers>...)]
+falkor-commander [(--s <scope>) | (-p <path>)] [(--k <keyword>)] [<tasks>...] [(-- <answers>...)]
 falkor-commander (-v | --version | -h | --help)
 ```
 
@@ -28,6 +28,9 @@ The **falkor-commander** project is a standalone npm command-line application wr
 
 `-s <scope>`, `--scope <scope>`
 :   The scope to look for plugins under **node_modules** (default: **@falkor**)
+
+`-p <path>`, `--path <path>`
+:   Explicit directory to look for plugins in (overrides scope setting)
 
 `-k <keyword>`, `--keyword <keyword>`
 :   The keyword to look for in plugin candidates' **package.json** (default: **@falkor-plugin**)
