+++ readme.md
@@ -26,8 +26,8 @@ Since both the workflow and API are subjects to change, if you want to follow th
 Usage:
 
 ```
-falkor-commander [(--scope <scope>)] [(--keyword <keyword>)] [<tasks>...] [(-- <answers>...)]
-falkor-commander [(--s <scope>)] [(--k <keyword>)] [<tasks>...] [(-- <answers>...)]
+falkor-commander [(--scope <scope>) | (--path <path>)] [(--keyword <keyword>)] [<tasks>...] [(-- <answers>...)]
+falkor-commander [(--s <scope>) | (-p <path>)] [(--k <keyword>)] [<tasks>...] [(-- <answers>...)]
 falkor-commander (-v | --version | -h | --help)
 ```
 
@@ -36,6 +36,7 @@ Options:
 * `-v` or `--version`: Show version and exit
 * `-h` or `--help`: Show help and exit
 * `-s <scope>` or `--scope <scope>`: The scope to look for plugins under `node_modules` (default: `@falkor`)
+* `-p <path>` or `--path <path>`: Explicit directory to look for plugins in (overrides scope setting)
 * `-k <keyword>` or `--keyword <keyword>`: The keyword to look for in plugin candidates' `package.json` (default: `@falkor-plugin`)
 * `<tasks>...`: Treat all positional arguments as buffered task IDs
 * `-- <answers>...`: Treat all positional arguments after double dash as buffered input
