+++ src/cli/index-cli.ts
@@ -14,14 +14,15 @@ version ${getOwnVersion()}
 Falkor plugin based task runner / -sequencer - part of the Falkor Framework
 
 Usage:
-  falkor-commander [(--scope <scope>)] [(--keyword <keyword>)] [<tasks>...] [(-- <answers>...)]
-  falkor-commander [(--s <scope>)] [(--k <keyword>)] [<tasks>...] [(-- <answers>...)]
+  falkor-commander [(--scope <scope>) | (--path <path>)] [(--keyword <keyword>)] [<tasks>...] [(-- <answers>...)]
+  falkor-commander [(--s <scope>) | (-p <path>)] [(--k <keyword>)] [<tasks>...] [(-- <answers>...)]
   falkor-commander (-v | --version | -h | --help)
 
 Options:
   -v, --version                      Show version and exit
   -h, --help                         Show this screen and exit
   -s <scope>, --scope <scope>        The scope to look for plugins under 'node_modules' [default: @falkor]
+  -p <path>, --path <path>           Explicit directory to look for plugins in (overrides scope setting)
   -k <keyword>, --keyword <keyword>  The keyword to look for in plugin candidates' 'package.json' [default: @falkor-plugin]
   <tasks>...                         Treat all positional arguments as buffered task IDs
   -- <answers>...                    Treat all positional arguments after double dash as buffered input
