+++ src/commander/FalkorCommander.ts
@@ -1,29 +1,13 @@
-import { posix as path } from "path";
-import { pathToFileURL } from "url";
-
 import minimist from "minimist";
-import { TaskRunner, Task, util as falkorUtil, FalkorError } from "@falkor/falkor-library";
-
-type PluginDescriptor = {
-    name: string;
-    version: string;
-    module: string;
-};
+import { util as falkorUtil, FalkorError } from "@falkor/falkor-library";
+import PluginTaskRunner from "../task/PluginTaskRunner.js";
 
 const enum FalkorCommanderErrorCodes {
     MISSING_BUFFERED_TASK = "commander-buffered-error"
 }
 
-export default class FalkorCommander extends TaskRunner {
-    private readonly spaceRe = / /g;
-    private readonly argvReplacerRe = /^:. /;
-
-    protected scope: string;
-    protected keyword: string;
+export default class FalkorCommander extends PluginTaskRunner {
     protected startTime: [number, number];
-    protected taskBuffer: string[];
-    protected initArgv: { [key: string]: any };
-    protected pluginArgv: { [key: string]: minimist.ParsedArgs } = {};
 
     constructor(argv: minimist.ParsedArgs) {
         super("Commander", argv["--"]?.length ? argv["--"] : null);
@@ -36,6 +20,9 @@ export default class FalkorCommander extends TaskRunner {
         this.keyword = argv.k || argv.keyword || this.config.external?.commander?.keyword || "@falkor-plugin";
         delete argv.k;
         delete argv.keyword;
+        this.forcedPluginPath = argv.p || argv.path || this.config.external?.commander?.path || null;
+        delete argv.p;
+        delete argv.path;
         if (Object.keys(argv).length === 0) {
             argv = null;
         }
@@ -51,36 +38,6 @@ export default class FalkorCommander extends TaskRunner {
         this.main()
             .then(() => process.exit(0))
             .catch(() => process.exit(1));
-    }
-
-    public register(task: Task): void {
-        if (this.taskBuffer && !this.taskBuffer.includes(task.id)) {
-            this.logger.info(
-                `${this.theme.formatWarning(`${this.warningPrompt} skipped`)} task '${this.theme.formatDebug(
-                    task.id
-                )}' (not present in ${this.theme.formatBullet("TASK BUFFER")})`
-            );
-            return;
-        }
-        super.register(task);
-        if (this.initArgv) {
-            const safeTaskId = task.id.replace(this.spaceRe, "-").toLowerCase();
-            if (this.initArgv[safeTaskId]) {
-                let argvStr = this.initArgv[safeTaskId];
-                let replacer = "#";
-                if (this.argvReplacerRe.test(argvStr)) {
-                    replacer = argvStr[1];
-                    argvStr = argvStr.substr(3);
-                }
-                this.pluginArgv[task.id] = minimist(
-                    falkorUtil.cliTokenize(argvStr.replace(new RegExp("\\" + replacer, "g"), "-")),
-                    {
-                        "--": true
-                    }
-                );
-            }
-        }
-        this.logger.info(`${this.theme.formatSuccess("registered")} task '${this.theme.formatDebug(task.id)}'`);
     }
 
     protected handleError(error: Error): Error | FalkorError {
@@ -102,11 +59,12 @@ export default class FalkorCommander extends TaskRunner {
     }
 
     protected async main(): Promise<void> {
-        this.startSubtask("Initialization");
-        await this.initPlugins();
-        this.endSubtaskSuccess("done");
+        await super.main();
 
-        if (this.taskBuffer) {
+        if (this.singlePluginMode) {
+            this.logger.info(`${this.infoPrompt} starting single plugin mode`);
+            return this.run(null, this.pluginArgv);
+        } else if (this.taskBuffer) {
             const missingTasks = this.taskBuffer.filter((id) => !this.collection[id]);
             if (missingTasks.length) {
                 this.logger.error(
@@ -167,106 +125,5 @@ export default class FalkorCommander extends TaskRunner {
             process.exit(0);
         }
         return selection;
-    }
-
-    protected async initPlugins(): Promise<void> {
-        this.logger.notice(`testing if 'cwd' is plugin (${this.theme.formatPath(this.cwd)})`);
-        const pluginTestPkg = this.testPackage(this.cwd);
-        if (pluginTestPkg) {
-            this.logger.info(`'cwd' is plugin, running in single-plugin test mode`);
-            await this.testPlugin(pluginTestPkg);
-            return;
-        }
-
-        this.logger.notice(
-            `'cwd' is not plugin, discovering scope '${this.theme.formatPath(
-                this.scope
-            )}' with required keyword '${this.theme.formatPath(this.keyword)}'`
-        );
-
-        for (const descriptor of this.testModule(this.cwd)) {
-            await this.testPlugin(descriptor);
-        }
-
-        delete this.initArgv;
-    }
-
-    protected testPackage(dir: string): PluginDescriptor {
-        const pkgPath = path.join(dir, "package.json");
-        if (this.shell.test("-f", pkgPath)) {
-            try {
-                const pkg = JSON.parse(this.shell.cat(pkgPath));
-                if (
-                    pkg.keywords &&
-                    Array.isArray(pkg.keywords) &&
-                    pkg.keywords.some((k: string) => k === this.keyword)
-                ) {
-                    return {
-                        name: pkg.name,
-                        version: pkg.version,
-                        module: pathToFileURL(path.join(dir, pkg.module)).toString()
-                    };
-                }
-            } catch (e) {}
-        }
-        return null;
-    }
-
-    protected testModule(dir: string): PluginDescriptor[] {
-        const ret: PluginDescriptor[] = [];
-        let prev: string;
-        do {
-            const modulePath = path.join(dir, "node_modules", this.scope);
-            this.logger.debug(`testing module path ${this.theme.formatPath(modulePath)}`);
-            if (this.shell.test("-d", modulePath)) {
-                this.shell.ls("-AL", modulePath).forEach((d) => {
-                    const testPath = path.join(modulePath, d);
-                    this.logger.notice(`testing module directory ${this.theme.formatPath(testPath)}`);
-                    const candidate = this.testPackage(testPath);
-                    if (candidate) {
-                        ret.push(candidate);
-                    }
-                });
-            }
-            prev = dir;
-            dir = path.join(dir, "..");
-        } while (prev !== dir && dir !== ".");
-        return ret;
-    }
-
-    protected async testPlugin(descriptor: PluginDescriptor): Promise<void> {
-        this.logger.notice(`importing module '${this.theme.formatDebug(descriptor.name)}' (${descriptor.version})`);
-        let moduleExports;
-        try {
-            moduleExports = (await import(descriptor.module)).default;
-        } catch (e) {
-            this.logger.warning(`failed to import module '${this.theme.formatDebug(descriptor.name)}'`).debug(e);
-            return;
-        }
-        if (!Array.isArray(moduleExports)) {
-            moduleExports = [moduleExports];
-        }
-        moduleExports.forEach((item) => {
-            if (item instanceof Task) {
-                this.register(item);
-                return;
-            }
-            //#if _DEBUG
-            // NOTE: instanceof check can cause trouble while developing with linked local packages, they may have their
-            // separate library installations on disk, since symlinked modules' dependencies do not get deduped
-            if (Object.prototype.toString.call(item) === "[object @FalkorTask]") {
-                this.logger.warning(
-                    `assuming '${this.theme.formatDebug(item.id)}' is Falkor task, but not instance of used library`
-                );
-                try {
-                    this.register(item);
-                    return;
-                } catch (e) {}
-            }
-            //#endif
-            this.logger.debug(
-                `failed to process item of '${this.theme.formatError(falkorUtil.getClassChain(item).join(" < "))}'`
-            );
-        });
     }
 }
